Rewrite basically everything
Picked this project back up after a year or more and ran into issues
thinking through how to handle the transport types outside of the mock.
The solution I landed on was to just rewrite the whole thing to be
async. Additionally I've actually written up the failure detection
and passive shuffling. The core join flow works as shown by the tests,
but I need to work on a more indepth test harness for all behaviors.