Refine failure handling algorithm, implement broadcast api
The "public" API for broadcasting a message to the network has been implemented.
Additionally while writing a test for failure handling I came across a pathalogical case
where all peers in a passive view reject the low priority neighbor
request. This means that failed peer replacement can stall. To address
this problem I've added an extra condition where if we've failed to find
a replacement after iterating over the entire passive view, we bump the
priority to High.