Test helper refactoring + Swap to LruCache in ConnectionManager impls
I moved the convergance loop out of the tests into its own function to
make things a bit cleaner. At the same time I switched from tracking
cached tonic clients in a HashMap to an LruCache.

This should ensure an upper bound on the number of active grpc clients
and better bound memory and fd resources.