Switch MessageStore to use LruCache instead of IndexSet
I totally misunderstood how index/pop worked for IndexSet.
I ran an overnight stress test in the homelab and noticed that after
about an hour of operation everything fell over and thousands of
messages were being forwarded when they should have been ignored as
already seen.

This made me super suspicious of the MessageStore behavior. I wrote a
simple test to make sure IndexSet::pop was actually doing what I
thought, and lo! it was not. In retrospect I was dumb for expecting LRU
like behavior. It adds new items to the end, and pop removes the last
element.

This meant that after we filled the MessageStore all new messages would
immediately pop the last messages instead of the oldest message. This
diff fixes that by using an actual LRU cache.